componentize talking points section
